+++ src/core/reply.ts
@@ -8,7 +8,7 @@ import handleTextbox from "./questions/textbox";
 import handleClipboard from "./questions/clipboard";
 import handleNumber from "./questions/number";
 import handleContentEditable from "./questions/contenteditable";
-import { injectionFunction } from "./code-listener";
+import { removeListener } from "./code-listener";
 
 /**
  * Reply to the question
@@ -29,41 +29,51 @@ async function reply(
   const question = createQuestion(config, form);
   const inputList: NodeListOf<HTMLElement> = form.querySelectorAll(query);
 
-  const response = await getChatGPTResponse(config, question).catch(
+  const gptAnswer = await getChatGPTResponse(config, question).catch(
     (error) => ({
       error,
     })
   );
 
-  const haveError = typeof response === "object" && "error" in response;
+  const haveError = typeof gptAnswer === "object" && "error" in gptAnswer;
 
   if (config.cursor)
     hiddenButton.style.cursor =
       config.infinite || haveError ? "pointer" : "initial";
 
   if (haveError) {
-    console.error(response.error);
+    console.error(gptAnswer.error);
     return;
   }
 
   if (config.logs) {
     Logs.question(question);
-    Logs.response(response);
+    Logs.response("Original: " + gptAnswer.response);
+    Logs.response("Normalized: " + gptAnswer.normalizedResponse);
   }
 
+  /* Handle clipboard mode */
   if (config.mode === "clipboard") {
-    return handleClipboard(config, response);
+    if (!config.infinite) removeListener(hiddenButton);
+    return handleClipboard(config, gptAnswer);
   }
 
+  /* Handle question to answer mode */
   if (config.mode === "question-to-answer") {
+    removeListener(hiddenButton);
+
     const questionBackup = form.textContent;
-    const questionContainer = form.querySelector(".qtext");
-    questionContainer.textContent = response;
+    const questionContainer = form.querySelector<HTMLElement>(".qtext");
+
+    questionContainer.textContent = gptAnswer.response;
+    questionContainer.style.whiteSpace = "pre-wrap";
+
     questionContainer.addEventListener("click", function () {
-      questionContainer.textContent =
-        questionContainer.textContent === questionBackup
-          ? response
-          : questionBackup;
+      const isNotResponse = questionContainer.textContent === questionBackup;
+      questionContainer.style.whiteSpace = isNotResponse ? "pre-wrap" : null;
+      questionContainer.textContent = isNotResponse
+        ? gptAnswer.response
+        : questionBackup;
     });
     return;
   }
@@ -77,15 +87,14 @@ async function reply(
   ];
 
   for (const handler of handlers) {
-    if (handler(config, inputList, response)) return;
+    if (handler(config, inputList, gptAnswer)) return;
   }
 
   /* In the case we can't auto complete the question */
-  handleClipboard(config, response);
+  handleClipboard(config, gptAnswer);
 
   /* Better then set once on the event because if there is an error the user can click an other time on the question */
-  if (!config.infinite)
-    hiddenButton.removeEventListener("click", injectionFunction);
+  if (!config.infinite) removeListener(hiddenButton);
 }
 
 export default reply;