+++ src/types/gptAnswer.ts
@@ -1,0 +1,6 @@
+type GPTAnswer = {
+  response: string;
+  normalizedResponse: string;
+};
+
+export default GPTAnswer;